Making 3D Models
Hello everyone,
Those days I didn't worked as I'm used too because I'm working on the 3D print prototype device.
When the device will be what I want it to be, I will publish it on the repo but for now I need clearly to fix some part of it.

For now at this state, the prototype is divided by two part : The Front Pannel and Rear Pannel.

However, I worked a bit on the source code; more on the UI but fixing some minors errors and some small details has been add.
I'm currently working on a way to display the quality of the Wi-Fi link in the UI but for now, I don't really know how to do it.

Stay tuned ,
ROCHAT Franck :night_with_stars:  .