Working on Background Video Loop
Hello Everyone,
Those days I worked on a way to clothe on the design of the app by adding a Video Loop in the background (At the begining it's was for the weather forecast).

For now , it's not working well I as I expected.
On Raspberry pi It's lagging as hell and on regular computer, the video got stopped every seconde because of the other Instruction getting called.
So Im working on 'Multiprocessing' to make it works but It's doesn't sound good for now hahahah  .

:D :+1:

Thank you,
Franck. :bridge_at_night: .