Update late Work
Hello evryone ,
Lately this month I worked on the consumption of fuel but I didn't integrate the code in the UI for now because now I want to keep focus on the feature that are already integrated in the project.

The other idea that aren't for now implemented will be integrated later but for now it's time to work on the DOC and the main program.

I had also worked on a way to dynamize the background and I think that after all test that I run that put a Video on kivy with the kind of program that i've made isn't a great idea for many reason :
1- It's make 'laging' small computers as RPI, so it's makes ugly and unusable UI.
2- Python don't use all the cores availables in the UC as I want so It's why I don't want to continue to use video in this kind of case. Maybe in the case where an Kivy app doesn't need to reach info on the web all the time, it's can work but not in our case.

So to resolve this problem, I think I will only put a simple JPG to the background (or a GIF, if it's works).

So now, I'm back and I will work on the doc and new project will comes up soon on my Github.

Thank You,
Franck. :night_with_stars: :love_hotel:  .