Kivy is Laggy ? Yes, it is.
Hello eevryone ,

Today i'm making a small update to integrate an old function that shows the emergencies numbers that you can call where-ever you're but it's not the main topic of this small update.

Today, I want to say that I will stop for a moment to continue to dev this program.
Why ?
Because Kivy is Laggy and its really annoying to see it as a final product.
For now, I don't know how to optimize it.
So i will take sometimes to work on it AND also find another UI.

To be honest, it's really frustrating to see that our works not always follow our plans as usual!

So I will work by my side but I will not give up for now !

Franck ROCHAT , :four_leaf_clover: :sunny:  .