More Ui , More Knowledge
More Ui , More Knowledge...

So for me the next step is to learn How to deploy a program fully.
And I learned it ; but I encountered some problems... of course .

With the TK UI everythings fine and works great.

However isn't the case with the Kivy UI.
Some problem occured on raspberry pi plateform with GL Drivers.
 --> More details here : https://github.com/Franck1333/FR_Template/releases/tag/1.13.1

So for now, I'm looking for a new UI that will replace the Kivy UI and can be deploy as a .deb package like Tkinter do.

Thank you all,
M.ROCHAT Franck. :+1:  :100:  :sunrise: :love_hotel: