Welcome to Flask
Hello everyone,

I've found a new framework which I've got a good feeling with it.

To get more freedom of 'design' and style, I want to try to make an UI in a web browser ; It's will be funny to dev and use it.

So those days, I will work on Flask and try to find out if this framework is usable in this project or not. ^^

That's all.
Thank you, Franck :sunny: :+1: :night_with_stars:  .