Back to 'Source'
Go back 404!!!

So, as I said later, I will stop to work on the WebInterface until I found a solution to fix the problem that I encoutered.

Anyway, for the moment I wanna come back to Tkinter and continue the project as it should be.

So, I re-introduce old feature and adjust them to be used in this project.

That's all for now,
Franck.