New feature
Hello everyone,

A new feature comes up and it's ready from now!

This feature called "Recherche par Adresse" allow you to search a location with a postal address parts and get the result as a map in the UI.

The map will be download inside the project folder at each execution of the function.

That's it for now...
Thank you, Franck :+1: :sunflower: